# web/swagger.yml
swagger: "2.0"
info:
  title: AI Camera
  description: API Description
  version: "1.0"
basePath: /
schemes:
  - http

paths:      
  /login:
    post:
      tags:
        - "0. Login"

      parameters:
        - name: loginfo
          in: body
          schema:
            type: object
            properties:
              user_id:
                type: string
                example: "user"
              password:
                type: string
                example: "0000"
          
      responses:
        200:
          description: success
  
  /logout:
    get:
      tags:
        - "0. Login"

      responses:
        200:
          description: success
  
  /get_info:
    get:
      summary: Get system information.
      tags:
        - "1. GET"

      produces:
        - application/json

      responses:
        200:
          description: None

  /service/{service}:
    get:
      tags:
        - "1. GET"

      parameters:
        - name: service
          in: path
          type: string
          description: get service current status

      responses:
        200:
          description: success
          
    post:
      summary: Activate/deactivate the service
      tags:
        - "2. POST"

      parameters:
        - name: service
          in: path
          type: string
          description: service name

        - name: command
          in: body
          schema:
            type: object
            properties:
              command:
                type: string
                enum: ["start", "stop"]
                example: start
          description: "Valid command: ['start', 'stop']"

      responses:
        200:
          description: success

  /get_image:
    get:
      tags:
        - "1. GET"

      parameters:
        - name: demo
          in: query
          type: string
          description: type something to get demo image
      
      responses:
        200:
          description: None

  /update_box_info:
    post:
      tags:
        - "3. Person tracking"

      parameters:
        - name: box information
          in: body
          schema:
            type: object
            properties:
              boxName:
                type: object
                properties:
                  x:
                    type: integer
                  y:
                    type: integer
                  w:
                    type: integer
                  h:
                    type: integer
                  newName:
                    type: string

      responses:
        200:
          description: None
  
  /get_box_info:
    get:
      summary: Get current box infomation
      tags:
        - "3. Person tracking"
      
      responses:
        200:
          schema:
              type: object
              properties:
                boxName:
                  type: object
                  properties:
                    x:
                      type: integer
                    y:
                      type: integer
                    w:
                      type: integer
                    h:
                      type: integer

        